migration/requirements.txt
firebase-admin==6.5.0
pymysql==1.1.0
python-dotenv==1.0.1
